更新官方库编码器驱动，修复bug，完成电机角度控制pid

+++ ebox_stm32f103RCT6_VS/ebox_stm32f103RCT6_VS/common/edriver/encoder_exti.cpp
@@ -51,7 +51,7 @@ void EncoderExti::begin()
 	extiB.interrupt(ENABLE);
 }
 
-short EncoderExti::getPos()
+long EncoderExti::getPos()
 {
 	return pos;
 }
@@ -67,7 +67,7 @@ void EncoderExti::refreshDiff()
 	posOld = pos;
 }
 
-short EncoderExti::getDiff()
+long EncoderExti::getDiff()
 {
 	return diff;
 }
